consolidate all botanix consensus checks to AuthorityConsensus
