+++ crates/consensus/authority/src/lib.rs
@@ -22,8 +22,10 @@
 use reth_botanix_lib::extra_data_header::ExtraDataHeader;
 use reth_consensus::{Consensus, ConsensusError};
 use reth_consensus_common::{
-    utils::{get_block_producer_address, unix_timestamp},
-    validation::{self, validate_poa_header_standalone},
+    utils::{
+        create_authority_sighash, get_block_producer_address, unix_timestamp, validate_inturn,
+    },
+    validation::validate_header_standalone,
 };
 use reth_interfaces::{
     executor::{BlockExecutionError, BlockValidationError},
@@ -48,9 +50,9 @@ use reth_revm::{
     processor::EVMProcessor, State,
 };
 use std::sync::Arc;
-use voting::{AuthorityVoteCollection, Vote};
-
 use tokio::sync::{RwLock, RwLockReadGuard, RwLockWriteGuard};
+use tracing::error;
+use voting::{AuthorityVoteCollection, Vote};
 
 use tracing::{trace, warn};
 mod block_builder;
@@ -72,7 +74,7 @@ pub use builder::AuthorityConsensusBuilder;
 /// Ethereum authority consensus
 ///
 /// This consensus engine does basic checks as outlined in the execution specs.
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub struct AuthorityConsensus {
     /// Configuration
     chain_spec: Arc<ChainSpec>,
@@ -88,7 +90,7 @@ impl AuthorityConsensus {
 
 impl Consensus for AuthorityConsensus {
     fn validate_header(&self, header: &SealedHeader) -> Result<(), ConsensusError> {
-        validation::validate_header_standalone(header, &self.chain_spec)?;
+        reth_consensus_common::validation::validate_header_standalone(header, &self.chain_spec)?;
         Ok(())
     }
 
@@ -107,7 +109,7 @@ impl Consensus for AuthorityConsensus {
     }
 
     fn validate_block(&self, block: &SealedBlock) -> Result<(), ConsensusError> {
-        validation::validate_block_standalone(block, &self.chain_spec)
+        reth_consensus_common::validation::validate_block_standalone(block, &self.chain_spec)
     }
 
     fn validate_header_with_total_difficulty(
@@ -115,7 +117,71 @@ impl Consensus for AuthorityConsensus {
         header: &Header,
         total_difficulty: U256,
     ) -> Result<(), ConsensusError> {
-        validation::validate_header_with_total_difficulty(header, total_difficulty)?;
+        reth_consensus_common::validation::validate_header_with_total_difficulty(
+            header,
+            total_difficulty,
+        )?;
+        Ok(())
+    }
+
+    /// Validate poa extra data header
+    fn validate_extra_data_header(
+        &self,
+        header: &Header,
+        authority_signers: &[secp256k1::PublicKey],
+    ) -> Result<(), ConsensusError> {
+        // Skip over genesis
+        if header.number == 0 {
+            return Ok(());
+        }
+        // First run the basic validation
+        reth_consensus_common::validation::validate_header_extradata(header)?;
+
+        // Attempt to deserialize the extra data header
+        let extra_data = reth_botanix_lib::extra_data_header::ExtraDataHeader::deserialize(
+            &mut header.extra_data.to_vec().as_slice(),
+        )
+        .map_err(|e| {
+            error!("Failed to deserialize extra data header: {:?}", e);
+            ConsensusError::ExtraDataInvalid
+        })?;
+        // Validate the authority signature and signature came from one of the authorities
+        let sig_hash = create_authority_sighash(&mut header.clone(), &extra_data);
+        extra_data.validate_authority_signature(&sig_hash.to_vec(), authority_signers).map_err(
+            |e| {
+                error!("Failed to validate authority signature: {:?}", e);
+                ConsensusError::InvalidAuthoritySignature
+            },
+        )?;
+        // TODO (armins) in the future this is where we would validate federation votes
+
+        Ok(())
+    }
+
+    /// Validate poa block beneficiary
+    fn validate_block_beneficiary(&self, header: &Header) -> Result<(), ConsensusError> {
+        if header.beneficiary != Address::ZERO {
+            return Err(ConsensusError::BlockBeneficiaryIsNotBurnAddress);
+        }
+
+        Ok(())
+    }
+
+    /// Validates PoA header standalone according to the authority consensus rules.
+    fn validate_header_standalone(
+        &self,
+        header: &Header,
+        authority_signers: &[secp256k1::PublicKey],
+    ) -> Result<(), ConsensusError> {
+        // Validate EDH serialization and signature on block
+        self.validate_extra_data_header(header, authority_signers)?;
+
+        // Validate fee benificiary
+        self.validate_block_beneficiary(header)?;
+
+        // Validate signer is in turn
+        validate_inturn(header, authority_signers)?;
+
         Ok(())
     }
 }
@@ -483,6 +549,7 @@ where
         authority_signers: &Vec<secp256k1::PublicKey>,
         witness_data: &Option<Vec<bitcoin::witness::Witness>>,
         utxo_commitment: &[u8; 32],
+        consensus: &AuthorityConsensus,
     ) -> Result<SealedHeader, BlockExecutionError> {
         let Block { header, body, .. } = block;
         let body = BlockBody { transactions: body, ommers: vec![], withdrawals: None };
@@ -503,7 +570,7 @@ where
         )?;
 
         // Redundant check. Lets make sure the header is valid
-        validate_poa_header_standalone(&header, authority_signers).map_err(|e| {
+        consensus.validate_header_standalone(&header, authority_signers).map_err(|e| {
             warn!(target: "consensus::authority", "failed to validate POA header: {:?}", e);
             // TODO(armins) return more expressive error
             BlockExecutionError::Validation(BlockValidationError::InvalidExtraData)
@@ -530,7 +597,7 @@ where
     // Execute and run poa validation on the block without inserting it into the storage
     pub(crate) fn execute_imported_block<EvmConfig>(
         &mut self,
-        chain_spec: Arc<ChainSpec>,
+        consensus: &AuthorityConsensus,
         sealed_block: SealedBlock,
         botanix_consensus_pkg: Option<BotanixConsensusPackage>,
         evm_config: EvmConfig,
@@ -547,7 +614,8 @@ where
             )))
             .with_bundle_update()
             .build();
-        let mut executor = EVMProcessor::new_with_state(chain_spec.clone(), db, evm_config);
+        let mut executor =
+            EVMProcessor::new_with_state(consensus.chain_spec.clone(), db, evm_config);
 
         let senders =
             TransactionSigned::recover_signers(&sealed_block.body, sealed_block.body.len()).ok_or(
@@ -560,13 +628,13 @@ where
 
         // validate before executing block
         let authority_signers = self.authorities.clone();
-        validate_poa_header_standalone(&sealed_block.header.clone(), &authority_signers).map_err(
-            |e| {
+        consensus
+            .validate_header_standalone(&sealed_block.header.clone(), &authority_signers)
+            .map_err(|e| {
                 warn!(target: "consensus::authority", "failed to validate POA header: {:?}", e);
                 // TODO(armins) return more expressive error
                 BlockExecutionError::Validation(BlockValidationError::InvalidExtraData)
-            },
-        )?;
+            })?;
 
         let block_builder_address = get_block_producer_address(&sealed_block.header.clone());
         let (bundle_state, _gas_used) = self.execute(
@@ -578,5 +646,394 @@ where
         )?;
 
         Ok(bundle_state)
+    }
+}
+
+
+#[cfg(test)]
+mod tests {
+    use std::str::FromStr;
+
+    use reth_consensus_common::utils::{block_fees_split, get_authority_list, get_in_turn_interval, is_inturn, recovery_authority, validate_against_parent, CoordinatorInterval};
+    use reth_primitives::BOTANIX_TESTNET;
+
+    use super::*;
+
+    #[allow(dead_code)]
+    const EDH_DEFAULT_SIGHASH: &str =
+        "0xaaa3492fe3eec8da1ca35aca5930a44b1a5805e813bdd1773678b5041d905276";
+
+    #[allow(dead_code)]
+    const SK1: &str = "1aabc5cc52b62b570dc69001f1ab49cd1a7056bf6312fe058f094135f2c9b019";
+    #[allow(dead_code)]
+    const SK2: &str = "1bc1f5cc52b62b570dc69001f1ab49cd1a7056bf6312fe058f094135f2c9b019";
+
+    #[allow(dead_code)]
+    fn generate_secret_key(hex_string: &str) -> secp256k1::SecretKey {
+        secp256k1::SecretKey::from_str(hex_string).expect("Invalid hex string for SecretKey")
+    }
+    #[allow(dead_code)]
+    fn sign_block_helper(header: &mut Header, signer_sk: Option<&str>) {
+        let mut edh = ExtraDataHeader::default();
+        let sk1 = generate_secret_key(SK1);
+        let sk2 = generate_secret_key(SK2);
+
+        edh.authority_signers = Some(vec![
+            secp256k1::PublicKey::from_secret_key(&secp256k1::Secp256k1::new(), &sk1),
+            secp256k1::PublicKey::from_secret_key(&secp256k1::Secp256k1::new(), &sk2),
+        ]);
+
+        let secp = secp256k1::Secp256k1::new();
+        header.number = 1;
+
+        let sighash = create_authority_sighash(&mut header.clone(), &edh);
+        let message = secp256k1::Message::from_slice(sighash.as_slice()).unwrap();
+        let signature = {
+            if let Some(sk_str) = signer_sk {
+                let sk = generate_secret_key(sk_str);
+                secp256k1::Secp256k1::sign_ecdsa_recoverable(&secp, &message, &sk)
+            } else {
+                // By default sign with the first authority
+                secp256k1::Secp256k1::sign_ecdsa_recoverable(&secp, &message, &sk1)
+            }
+        };
+        edh.set_signature(signature);
+
+        header.extra_data = Bytes::from(edh.serialize());
+    }
+
+    /* Tests for create authority sighash utility */
+    #[test]
+    fn create_default_edh_sighhash() {
+        let edh = ExtraDataHeader::default();
+        let mut header = Header::default();
+        let sighash = create_authority_sighash(&mut header, &edh);
+
+        assert_eq!(sighash.to_string(), EDH_DEFAULT_SIGHASH);
+    }
+
+    #[test]
+    fn create_sighash_with_authority_signature() {
+        // regarless of the signature, the sighash should be the same
+        // This is because we remove the signature from the extra data header before signing
+        let mut edh = ExtraDataHeader::default();
+        edh.set_signature(
+            secp256k1::ecdsa::RecoverableSignature::from_compact(
+                &[0u8; 64],
+                secp256k1::ecdsa::RecoveryId::from_i32(1i32).unwrap(),
+            )
+            .unwrap(),
+        );
+        let mut header = Header::default();
+        let sighash = create_authority_sighash(&mut header, &edh);
+
+        assert_eq!(sighash.to_string(), EDH_DEFAULT_SIGHASH);
+    }
+    #[test]
+    fn create_sighash_with_authorities() {
+        // However adding something else such as authority members should result in a different
+        // sighash
+        let mut edh = ExtraDataHeader::default();
+        edh.authority_signers = Some(vec![
+            secp256k1::PublicKey::from_secret_key(
+                &secp256k1::Secp256k1::new(),
+                &secp256k1::SecretKey::from_str(
+                    "1aabc5cc52b62b570dc69001f1ab49cd1a7056bf6312fe058f094135f2c9b019",
+                )
+                .unwrap(),
+            ),
+            secp256k1::PublicKey::from_secret_key(
+                &secp256k1::Secp256k1::new(),
+                &secp256k1::SecretKey::from_str(
+                    "1bc1f5cc52b62b570dc69001f1ab49cd1a7056bf6312fe058f094135f2c9b019",
+                )
+                .unwrap(),
+            ),
+        ]);
+        let mut header = Header::default();
+        let sighash = create_authority_sighash(&mut header, &edh);
+
+        assert_ne!(sighash.to_string(), EDH_DEFAULT_SIGHASH);
+    }
+
+    // Get authority list tests
+    #[test]
+    fn should_recover_none_authorities() {
+        let edh = ExtraDataHeader::default();
+        let mut header = Header::default();
+        header.extra_data = Bytes::from(edh.serialize());
+        let signer_list = get_authority_list(&header).unwrap();
+
+        assert_eq!(signer_list, None);
+    }
+
+    #[test]
+    fn should_recovery_authorities() {
+        let mut header = Header::default();
+        sign_block_helper(&mut header, None);
+        let edh = ExtraDataHeader::deserialize(&mut header.extra_data.to_vec().as_slice()).unwrap();
+        let signer_list = get_authority_list(&header).unwrap();
+
+        assert_eq!(signer_list, edh.authority_signers);
+    }
+
+    #[test]
+    fn fails_to_recover_when_edh_invalid() {
+        let mut header = Header::default();
+        header.extra_data = Bytes::from("foobar");
+        let signer_list = get_authority_list(&header);
+
+        assert!(signer_list.is_err());
+    }
+
+    // Tests for recover authority pk
+    #[test]
+    fn should_recover_authority() {
+        let mut header = Header::default();
+        sign_block_helper(&mut header, None);
+        let edh = ExtraDataHeader::deserialize(&mut header.extra_data.to_vec().as_slice()).unwrap();
+
+        let recovered = recovery_authority(&header).unwrap();
+        // utility function above only signs with the first authority
+        assert_eq!(recovered, edh.authority_signers.unwrap()[0]);
+    }
+
+    // Tests for validating poa extra data header
+    #[test]
+    fn should_skip_over_genesis() {
+        let consensus = AuthorityConsensus::new(Arc::new(BOTANIX_TESTNET.as_ref().to_owned()));
+        let mut header = Header::default();
+        header.number = 0;
+        let authority_signers = vec![];
+        let result = consensus.validate_extra_data_header(&header, &authority_signers);
+
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn should_fail_on_invalid_signature() {
+        let consensus = AuthorityConsensus::new(Arc::new(BOTANIX_TESTNET.as_ref().to_owned()));
+        // In this case we are signing with a non federation different key
+        let mut edh = ExtraDataHeader::default();
+        let sk1 = secp256k1::SecretKey::from_str(
+            "1aabc5cc52b62b570dc69001f1ab49cd1a7056bf6312fe058f094135f2c9b019",
+        )
+        .unwrap();
+        let non_fed = secp256k1::SecretKey::from_str(
+            "1bc1f5cc52b62b570dc69001f1ab49cd1a7056bf6312fe058f094135f2c9b019",
+        )
+        .unwrap();
+
+        edh.authority_signers =
+            Some(vec![secp256k1::PublicKey::from_secret_key(&secp256k1::Secp256k1::new(), &sk1)]);
+
+        let secp = secp256k1::Secp256k1::new();
+        let mut header = Header::default();
+        header.number = 1;
+
+        let sighash = create_authority_sighash(&mut header, &edh);
+        let message = secp256k1::Message::from_digest_slice(sighash.as_slice()).unwrap();
+        let signature = secp256k1::Secp256k1::sign_ecdsa_recoverable(&secp, &message, &non_fed);
+
+        edh.set_signature(signature);
+
+        header.extra_data = Bytes::from(edh.serialize());
+        let authority_signers = vec![];
+        let result = consensus.validate_extra_data_header(&header, &authority_signers);
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn should_validate_poa_header() {
+        // In this case we are signing with a non federation different key
+        let consensus = AuthorityConsensus::new(Arc::new(BOTANIX_TESTNET.as_ref().to_owned()));
+        let mut header = Header::default();
+        sign_block_helper(&mut header, None);
+        let edh = ExtraDataHeader::deserialize(&mut header.extra_data.to_vec().as_slice()).unwrap();
+        let authority_signers = edh.authority_signers.unwrap();
+        let result = consensus.validate_extra_data_header(&header, &authority_signers);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn unix_timestamp() {
+        let timestamp = super::unix_timestamp();
+        assert!(timestamp > 0);
+    }
+
+    #[test]
+    fn should_validate_poa_block_beneficiary() {
+        // default beneficiary is the burn address
+        let consensus = AuthorityConsensus::new(Arc::new(BOTANIX_TESTNET.as_ref().to_owned()));
+        let header = Header::default();
+        let result = consensus.validate_block_beneficiary(&header);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn should_fail_validate_poa_block_beneficiary() {
+        let consensus = AuthorityConsensus::new(Arc::new(BOTANIX_TESTNET.as_ref().to_owned()));
+        let mut header = Header::default();
+        header.beneficiary =
+            Address::from_str("0x4e0f6e05C8ca4b3dc2B7b7Ad6249B149b1980394").unwrap();
+        let result = consensus.validate_block_beneficiary(&header);
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn validate_against_parent_skip_gensis() {
+        let mut parent = Header::default();
+        parent.number = 0;
+        let current = Header::default();
+        let result = validate_against_parent(parent, current);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn should_fail_with_same_signer() {
+        let mut parent = Header::default();
+        let mut current = Header::default();
+
+        parent.number = 1;
+        current.number = 2;
+
+        sign_block_helper(&mut parent, None);
+        sign_block_helper(&mut current, None);
+
+        let result = validate_against_parent(parent, current);
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn should_pass_after_sufficient_time() {
+        let mut parent = Header::default();
+        let mut current = Header::default();
+
+        parent.number = 1;
+        parent.timestamp = 1704834442_u64;
+        current.number = 2;
+        current.timestamp = 1704834442_u64 + 60;
+
+        sign_block_helper(&mut parent, None);
+        sign_block_helper(&mut current, None);
+
+        let result = validate_against_parent(parent, current);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn should_pass_with_different_signer() {
+        let mut parent = Header::default();
+        let mut current = Header::default();
+        parent.number = 1;
+        current.number = 2;
+
+        sign_block_helper(&mut parent, None);
+        sign_block_helper(&mut current, Some(SK2));
+
+        let result = validate_against_parent(parent, current);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn is_inturn_true() {
+        let authorities_len = 1;
+        let signer_index = 0;
+        assert!(is_inturn(authorities_len, signer_index));
+    }
+
+    #[test]
+    fn is_inturn_false() {
+        let authorities_len = 1;
+        let signer_index = 1;
+        assert!(!is_inturn(authorities_len, signer_index));
+    }
+
+    #[test]
+    fn validate_inturn_ok() {
+        let mut header = Header::default();
+        header.timestamp = 1705621229;
+        sign_block_helper(&mut header, Some(SK1));
+
+        assert!(validate_inturn(
+            &header,
+            &[
+                secp256k1::PublicKey::from_secret_key(
+                    &secp256k1::Secp256k1::new(),
+                    &secp256k1::SecretKey::from_str(SK1).unwrap(),
+                ),
+                secp256k1::PublicKey::from_secret_key(
+                    &secp256k1::Secp256k1::new(),
+                    &secp256k1::SecretKey::from_str(SK2).unwrap(),
+                ),
+            ]
+        )
+        .is_ok());
+
+        // Sign the same header with a different key should fail
+
+        sign_block_helper(&mut header, Some(SK2));
+
+        assert!(validate_inturn(
+            &header,
+            &[
+                secp256k1::PublicKey::from_secret_key(
+                    &secp256k1::Secp256k1::new(),
+                    &secp256k1::SecretKey::from_str(SK1).unwrap(),
+                ),
+                secp256k1::PublicKey::from_secret_key(
+                    &secp256k1::Secp256k1::new(),
+                    &secp256k1::SecretKey::from_str(SK2).unwrap(),
+                ),
+            ]
+        )
+        .is_err());
+    }
+
+    #[test]
+    fn should_split_rewards() {
+        let base_block_reward = 100;
+        let (botanix_reward, beneficiary_reward) = block_fees_split(base_block_reward);
+        assert_eq!(botanix_reward, 20);
+        assert_eq!(beneficiary_reward, 80);
+    }
+
+    #[test]
+    fn should_get_block_producer_address_from_header() {
+        let mut header = Header::default();
+        sign_block_helper(&mut header, None);
+        let edh = ExtraDataHeader::deserialize(&mut header.extra_data.to_vec().as_slice()).unwrap();
+        let block_producer_address = get_block_producer_address(&header);
+        assert_eq!(
+            block_producer_address,
+            public_key_to_address(edh.authority_signers.unwrap()[0])
+        );
+    }
+
+    #[test]
+    fn should_return_zero_address_when_no_authority() {
+        let header = Header::default();
+        let block_producer_address = get_block_producer_address(&header);
+        assert_eq!(block_producer_address, Address::ZERO);
+    }
+
+    #[test]
+    fn get_inturn_interval() {
+        let authorities_len = 10;
+        let signer_index = 3; // Example signer index
+        let (start, end, time_passed, time_remaining): CoordinatorInterval =
+            get_in_turn_interval(authorities_len, signer_index);
+        let current_ts = super::unix_timestamp();
+        println!(
+            "Signer index {} is in turn from {}s to {}s. Current ts = {:?}s. Time passed = {:?}s, time remaining = {:?}s",
+            signer_index,
+            start,
+            end,
+            super::unix_timestamp(),
+            time_passed,
+            time_remaining,
+        );
+        assert!(current_ts > start);
+        assert!(current_ts < end);
     }
 }
