+++ crates/consensus/auto-seal/Cargo.toml
@@ -32,6 +32,12 @@ tokio = { workspace = true, features = ["sync", "time"] }
 tokio-stream.workspace = true
 tracing.workspace = true
 
+secp256k1 = { workspace = true, features = [
+    "global-context",
+    "rand-std",
+    "recovery",
+] }
+
 [dev-dependencies]
 reth-interfaces = { workspace = true, features = ["test-utils"] }
 reth.workspace = true
