+++ crates/consensus/auto-seal/src/lib.rs
@@ -91,6 +91,26 @@ impl Consensus for AutoSealConsensus {
     fn validate_block(&self, _block: &SealedBlock) -> Result<(), ConsensusError> {
         Ok(())
     }
+
+    fn validate_extra_data_header(
+        &self,
+        _header: &Header,
+        _authority_signers: &[secp256k1::PublicKey],
+    ) -> Result<(), ConsensusError> {
+        Ok(())
+    }
+
+    fn validate_block_beneficiary(&self, _header: &Header) -> Result<(), ConsensusError> {
+        Ok(())
+    }
+
+    fn validate_header_standalone(
+        &self,
+        _header: &Header,
+        _authority_signers: &[secp256k1::PublicKey],
+    ) -> Result<(), ConsensusError> {
+        Ok(())
+    }
 }
 
 /// Builder type for configuring the setup
