+++ crates/consensus/beacon-core/Cargo.toml
@@ -16,5 +16,11 @@ reth-consensus-common.workspace = true
 reth-primitives.workspace = true
 reth-consensus.workspace = true
 
+secp256k1 = { workspace = true, features = [
+    "global-context",
+    "rand-std",
+    "recovery",
+] }
+
 [features]
 optimism = ["reth-primitives/optimism"]
