+++ crates/consensus/common/src/validation.rs
@@ -9,8 +9,6 @@ use reth_primitives::{
 };
 use reth_provider::{HeaderProvider, WithdrawalsProvider};
 
-use crate::utils;
-
 /// Validate header standalone
 pub fn validate_header_standalone(
     header: &SealedHeader,
@@ -202,22 +200,23 @@ pub fn validate_header_extradata(_header: &Header) -> Result<(), ConsensusError>
     Ok(())
 }
 
-/// Validates PoA header standalone according to the authority consensus rules.
-pub fn validate_poa_header_standalone(
-    header: &Header,
-    authority_signers: &[secp256k1::PublicKey],
-) -> Result<(), ConsensusError> {
-    // Validate EDH serialization and signature on block
-    utils::validate_poa_extra_data_header(header, authority_signers)?;
+// /// Validates PoA header standalone according to the authority consensus rules.
+// pub fn validate_poa_header_standalone(
+//     consensus: &AuthorityConsensus,
+//     header: &Header,
+//     authority_signers: &[secp256k1::PublicKey],
+// ) -> Result<(), ConsensusError> {
+//     // Validate EDH serialization and signature on block
+//     utils::validate_poa_extra_data_header(header, authority_signers)?;
 
-    // Validate fee benificiary
-    utils::validate_poa_block_beneficiary(header)?;
+//     // Validate fee benificiary
+//     utils::validate_poa_block_beneficiary(header)?;
 
-    // Validate signer is in turn
-    utils::validate_inturn(header, authority_signers)?;
+//     // Validate signer is in turn
+//     utils::validate_inturn(header, authority_signers)?;
 
-    Ok(())
-}
+//     Ok(())
+// }
 
 /// Validates the header with the given total difficulty.
 /// Note: Used to validate PoA extra data header
