+++ crates/consensus/consensus/Cargo.toml
@@ -13,6 +13,12 @@ workspace = true
 [dependencies]
 reth-primitives.workspace = true
 
+secp256k1 = { workspace = true, features = [
+    "global-context",
+    "rand-std",
+    "recovery",
+] }
+
 # misc
 auto_impl.workspace = true
 thiserror.workspace = true
