+++ crates/consensus/consensus/src/test_utils.rs
@@ -67,4 +67,36 @@ impl Consensus for TestConsensus {
             Ok(())
         }
     }
+
+    fn validate_extra_data_header(
+        &self,
+        _header: &Header,
+        _authority_signers: &[secp256k1::PublicKey],
+    ) -> Result<(), ConsensusError> {
+        if self.fail_validation() {
+            Err(ConsensusError::BaseFeeMissing)
+        } else {
+            Ok(())
+        }
+    }
+
+    fn validate_block_beneficiary(&self, _header: &Header) -> Result<(), ConsensusError> {
+        if self.fail_validation() {
+            Err(ConsensusError::BaseFeeMissing)
+        } else {
+            Ok(())
+        }
+    }
+
+    fn validate_header_standalone(
+        &self,
+        _header: &Header,
+        _authority_signers: &[secp256k1::PublicKey],
+    ) -> Result<(), ConsensusError> {
+        if self.fail_validation() {
+            Err(ConsensusError::BaseFeeMissing)
+        } else {
+            Ok(())
+        }
+    }
 }
